Bugfix: grouping taxonomies get reset in overviewss.

+++ app/src/Bolt/Storage.php
@@ -2580,6 +2580,14 @@ class Storage
                 // If it's like 'desktop#10', split it into value and sortorder..
                 list($slug, $sortorder) = explode('#', $slug . "#");
 
+                // @todo clean up and/or refactor
+                // If you save this content via anything other than the Bolt
+                // backend (see Content->setFromPost), then taxonomies that
+                // behave like groupings, will have their sortorders reset to 0.
+                if ($configTaxonomies[$taxonomytype]['behaves_like'] == 'grouping' && empty($sortorder) && $sortorder !== '0') {
+                    $sortorder = $currentsortorder;
+                }
+
                 if (empty($sortorder)) {
                     $sortorder = 0;
                 }
@@ -2615,6 +2623,7 @@ class Storage
                         'name' => $name,
                         'sortorder' => $sortorder
                     );
+
                     $this->app['db']->insert($tablename, $row);
                 }
 
@@ -2625,8 +2634,9 @@ class Storage
 
                 // Make it look like 'desktop#10'
                 $valuewithorder = $slug . "#" . $currentsortorder;
+                $slugkey = '/'.$configTaxonomies[$taxonomytype]['slug'].'/'.$slug;
 
-                if (!in_array($slug, $newslugs) && !in_array($valuewithorder, $newslugs)) {
+                if (!in_array($slug, $newslugs) && !in_array($valuewithorder, $newslugs) && !array_key_exists($slugkey, $newslugs)) {
                     $this->app['db']->delete($tablename, array('id' => $id));
                 }
             }
